Refactor listing invoices
